allow assigning unmapped json members to particular contract members
This allows json members that aren't mapped to a particular field in a
data contract to be assigned to a particular set of contract fields, one
for the json member's name and one for the json member's value.  With
this, the json-based hierarchical config source now has more readable
and compact config files.